+++ actix-multipart/CHANGES.md
@@ -2,7 +2,10 @@
 
 ## Unreleased
 
+- Add `MultipartForm` support for `Option<Vec<T>>` fields. [#3577]
 - Minimum supported Rust version (MSRV) is now 1.88.
+
+[#3577]: https://github.com/actix/actix-web/pull/3577
 
 ## 0.7.2